+++ src/main/java/com/newworld/saegil/configuration/PropertiesConfiguration.java
@@ -1,6 +1,6 @@
 package com.newworld.saegil.configuration;
 
-import com.newworld.saegil.proxy.config.ProxyProperties;
+import com.newworld.saegil.llm.config.ProxyProperties;
 import com.newworld.saegil.security.jwt.JwtProperties;
 import com.newworld.saegil.security.oauth2.KakaoOAuth2Properties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
