+++ src/main/java/com/newworld/saegil/llm/controller/ChatGptTextRequest.java
@@ -1,4 +1,4 @@
-package com.newworld.saegil.proxy.controller;
+package com.newworld.saegil.llm.controller;
 
 public record ChatGptTextRequest(
         String text
